Add azure functions for module twin
