+++ IoTDeviceSimulator/azure-functions/create-device/index.js
@@ -5,16 +5,22 @@
 
 module.exports = async function (context, req) {
     try {
+        console.log("Entering create-device");
         let res;
         let iothub = require('azure-iothub');
         let registry = iothub.Registry.fromConnectionString(process.env[req.body.connectionStringId]);
-        if(process.env[req.body.deviceInterfaceId]){
+        console.log(req.body.isDeviceTwin);
+        if(req.body.isDeviceTwin === 'true'){
             const deviceId = req.body.deviceInterfaceId;
             res = await registry.create({ deviceId});
+            console.log('create-device for deviceTwin ' + deviceId);
         }
         else{
+            console.log("Entering create-device for module twin ");
             const moduleId = req.body.deviceId;
-            res = await registry.create({ moduleId });
+            const deviceId = req.body.deviceInterfaceId;
+            res = await registry.addModule({ deviceId, moduleId });
+            console.log('create-device for moduleTwin' + deviceId + "," + moduleId);
         }
         return {
             body: res.responseBody
